Completed the data requests and most of the data answer framework

+++ src/main/java/com/bigb/sinope/request/DatalessRequest.java
@@ -34,7 +34,7 @@ public final class DatalessRequest extends AbstractRequest {
     }
 
     @Override
-    protected short getDataSize() {
+    protected int getDataSize() {
         return 0;
     }
